server/requirements/production.txt
Django==5.0.5
django-environ==0.11.2
drf-yasg==1.21.7
setuptools==69.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-filter==24.2
channels[daphne]==4.1.0
channels-redis==4.2.0
gunicorn==22.0.0
psycopg2==2.9.9
whitenoise==6.6.0
